Part 11

+++ domain/src/main/java/com/codewithfk/domain/usecase/DeleteProductUseCase.kt
@@ -4,5 +4,5 @@ import com.codewithfk.domain.model.CartItemModel
 import com.codewithfk.domain.repository.CartRepository
 
 class DeleteProductUseCase(private val cartRepository: CartRepository) {
-    suspend fun execute(cartItemId: Int, userId: Int) = cartRepository.deleteItem(cartItemId, userId)
+    suspend fun execute(cartItemId: Int, userId: Long) = cartRepository.deleteItem(cartItemId, userId)
 }
